user_service/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
cassandra-driver
pydantic[email]
